Make v2 endpoints the default for sessions
