+++ src/graphdatascience/session/session_lifecycle_manager.py
@@ -1,7 +1,20 @@
+from abc import ABC, abstractmethod
+
 from graphdatascience.session.aura_api import AuraApi
 
 
-class SessionLifecycleManager:
+class LifecycleManager(ABC):
+    @abstractmethod
+    def delete(self) -> bool:
+        pass
+
+    @abstractmethod
+    def verify_health(self) -> None:
+        # Raises a SessionStatusError if the season is in an unhealthy state
+        pass
+
+
+class SessionLifecycleManager(LifecycleManager):
     def __init__(self, session_id: str, aura_api: AuraApi):
         self.session_id = session_id
         self._aura_api = aura_api
@@ -12,3 +25,11 @@ class SessionLifecycleManager:
     def verify_health(self) -> None:
         # Raises a SessionStatusError if the season is in an unhealthy state
         self._aura_api.get_session(self.session_id)
+
+
+class Noop(LifecycleManager):
+    def delete(self) -> bool:
+        return True
+
+    def verify_health(self) -> None:
+        pass
