+++ tests/integrationV2/procedure_surface/session/test_session_api_spec_coverage.py
@@ -1,4 +1,4 @@
-from graphdatascience.session.session_v2_endpoints import SessionV2Endpoints
+from graphdatascience.session import AuraGraphDataScience
 from tests.integrationV2.procedure_surface.api_spec_coverage_test_helper import (
     assert_api_spec_coverage,
 )
@@ -6,4 +6,4 @@ from tests.integrationV2.procedure_surface.gds_api_spec import EndpointWithModes
 
 
 def test_session_api_spec_coverage(gds_api_spec: list[EndpointWithModesSpec]) -> None:
-    assert_api_spec_coverage(SessionV2Endpoints, gds_api_spec)
+    assert_api_spec_coverage(AuraGraphDataScience, gds_api_spec)
