+++ tests/unit/procedure_surface/arrow/test_triangles_arrow_endpoints.py
@@ -3,13 +3,15 @@ from unittest import mock
 import pandas as pd
 
 from graphdatascience.arrow_client.authenticated_flight_client import AuthenticatedArrowClient
-from graphdatascience.session.session_v2_endpoints import SessionV2Endpoints
+from graphdatascience.session import AuraGraphDataScience
+from graphdatascience.session.session_lifecycle_manager import Noop
 
 
 def test_session_exposes_top_level_triangles() -> None:
-    endpoints = SessionV2Endpoints(
+    endpoints = AuraGraphDataScience(
         mock.Mock(spec=AuthenticatedArrowClient),
-        db_client=None,
+        db_query_runner=None,
+        session_lifecycle_manager=Noop(),
         show_progress=False,
     )
