+++ tests/unit/session/test_aura_graph_data_science.py
@@ -1,24 +1,21 @@
+from pandas import DataFrame
 from pytest_mock import MockerFixture
 
 from graphdatascience import ServerVersion
-from graphdatascience.query_runner.neo4j_query_runner import Neo4jQueryRunner
+from graphdatascience.arrow_client.authenticated_flight_client import AuthenticatedArrowClient
 from graphdatascience.query_runner.query_mode import QueryMode
 from graphdatascience.session.aura_graph_data_science import AuraGraphDataScience
-from graphdatascience.session.session_lifecycle_manager import SessionLifecycleManager
-from graphdatascience.session.session_v2_endpoints import SessionV2Endpoints
+from graphdatascience.session.session_lifecycle_manager import Noop, SessionLifecycleManager
 from tests.unit.conftest import CollectingQueryRunner
 
 
 def test_run_cypher_write(mocker: MockerFixture) -> None:
     v = ServerVersion(9, 9, 9)
-    query_runner = CollectingQueryRunner(v)
+    query_runner = CollectingQueryRunner(v, {"version": DataFrame.from_dict({"version": ["v3"]})})
     gds = AuraGraphDataScience(
-        query_runner=query_runner,
-        db_query_runner=None,
-        session_lifecycle_manager=mocker.Mock(spec=SessionLifecycleManager),
-        gds_version=v,
-        v2_endpoints=mocker.Mock(),
-        authenticated_arrow_client=mocker.Mock(),
+        mocker.Mock(),
+        db_query_runner=query_runner,
+        session_lifecycle_manager=Noop(),
     )
 
     gds.run_cypher("RETURN 1", params={"foo": 1}, mode=QueryMode.WRITE, database="bar", retryable=True)
@@ -30,14 +27,11 @@ def test_run_cypher_write(mocker: MockerFixture) -> None:
 
 def test_run_cypher_read(mocker: MockerFixture) -> None:
     v = ServerVersion(9, 9, 9)
-    query_runner = CollectingQueryRunner(v)
+    query_runner = CollectingQueryRunner(v, {"version": DataFrame.from_dict({"version": ["v3"]})})
     gds = AuraGraphDataScience(
-        query_runner=query_runner,
-        db_query_runner=None,
-        session_lifecycle_manager=mocker.Mock(spec=SessionLifecycleManager),
-        gds_version=v,
-        v2_endpoints=mocker.Mock(),
-        authenticated_arrow_client=mocker.Mock(),
+        mocker.Mock(),
+        db_query_runner=query_runner,
+        session_lifecycle_manager=Noop(),
     )
 
     gds.run_cypher("RETURN 1", params={"foo": 1}, mode=QueryMode.READ, retryable=False)
@@ -54,37 +48,27 @@ def test_run_cypher_read(mocker: MockerFixture) -> None:
 
 
 def test_verify_connectivity(mocker: MockerFixture) -> None:
-    v = ServerVersion(9, 9, 9)
-    query_runner = mocker.Mock(spec=Neo4jQueryRunner)
+    arrow_client = mocker.Mock(spec=AuthenticatedArrowClient)
     session_lifecycle_manager = mocker.Mock(spec=SessionLifecycleManager)
-    v2_endpoints = mocker.Mock(spec=SessionV2Endpoints)
     gds = AuraGraphDataScience(
-        query_runner=query_runner,
-        db_query_runner=None,
-        session_lifecycle_manager=session_lifecycle_manager,
-        gds_version=v,
-        v2_endpoints=v2_endpoints,
-        authenticated_arrow_client=mocker.Mock(),
+        arrow_client,
+        None,
+        session_lifecycle_manager,
     )
 
     gds.verify_connectivity()
 
     session_lifecycle_manager.verify_health.assert_called_once()
-    v2_endpoints.verify_session_connectivity.assert_called_once()
-    v2_endpoints.verify_db_connectivity.assert_called_once()
+    arrow_client.request_token.assert_called_once()
 
 
 def test_delete(mocker: MockerFixture) -> None:
-    v = ServerVersion(9, 9, 9)
-    query_runner = CollectingQueryRunner(v)
+    arrow_client = mocker.Mock(spec=AuthenticatedArrowClient)
     session_lifecycle_manager = mocker.Mock(spec=SessionLifecycleManager)
     gds = AuraGraphDataScience(
-        query_runner=query_runner,
-        db_query_runner=query_runner,
-        session_lifecycle_manager=session_lifecycle_manager,
-        gds_version=v,
-        v2_endpoints=mocker.Mock(),
-        authenticated_arrow_client=mocker.Mock(),
+        arrow_client,
+        None,
+        session_lifecycle_manager,
     )
 
     gds.delete()
